fix: change some resource templates into resources
team_members, mob_session_coauthors, mob_session_coauthor_trailers and git_mob_version are static resources that do not depend on any input. Therefore these are changed from resource templates into resources. This makes them work in the MCP Inspector

Kept git_mob_help as a resource template as its dynamic - the content changes when command is supplied. This is yet to be tested to be working in MCP Inspector